% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tune_result.R
\name{clean_metric_name}
\alias{clean_metric_name}
\title{Clean Metric Names for Display}
\usage{
clean_metric_name(metric_name)
}
\arguments{
\item{metric_name}{Character, metric name with suffix}
}
\value{
Character, cleaned name
}
\description{
Clean Metric Names for Display
}
\keyword{internal}
