% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tune_result.R
\name{create_line_plot}
\alias{create_line_plot}
\title{Create Line Plot}
\usage{
create_line_plot(results, metric, tune_obj)
}
\arguments{
\item{results}{Results data frame}

\item{metric}{Metric to plot}

\item{tune_obj}{Original tune_result object}
}
\description{
Create Line Plot
}
\keyword{internal}
