% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_block_splsda.R
\name{create_results_dataframe}
\alias{create_results_dataframe}
\title{Create Results Data Frame}
\usage{
create_results_dataframe(results, param_combinations)
}
\arguments{
\item{results}{List of result objects}

\item{param_combinations}{Parameter combinations data frame}
}
\value{
Data frame with results
}
\description{
Converts list of results into a structured data frame.
}
\keyword{internal}
