% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tune_result.R
\name{create_scatter_plot}
\alias{create_scatter_plot}
\title{Create Scatter Plot}
\usage{
create_scatter_plot(results, metric, tune_obj)
}
\arguments{
\item{results}{Results data frame}

\item{metric}{Primary metric to plot}

\item{tune_obj}{Original tune_result object}
}
\description{
Create Scatter Plot
}
\keyword{internal}
