% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneR.r
\name{eda_tunr}
\alias{eda_tunr}
\title{eda_tunr : will calculate the following statistics on the datafiles
RMS, Kurtosis, Entropy, Skewness, crestfactor
Then it aloss calculates the Mahalanobis distance of these features between the different bearings to 
explore of there is significant correlation}
\usage{
eda_tunr(Result)
}
\arguments{
\item{Result}{is a list of raw vibration datafiles, the sample_size, the number of channels, and the dataset}
}
\value{
a list of datasets for each feature rms,kurtosis, entropy, creatfactor, skewness, and a flatfile will all plus the Mahalanobis   distances
}
\description{
eda_tunr : will calculate the following statistics on the datafiles
RMS, Kurtosis, Entropy, Skewness, crestfactor
Then it aloss calculates the Mahalanobis distance of these features between the different bearings to 
explore of there is significant correlation
}
\examples{

eda_tunr<-function(Result=list("filenames",channel_count,data_path,sample_size,dataset) 

}
