% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_block_splsda.R
\name{extract_keepX_for_combination}
\alias{extract_keepX_for_combination}
\title{Extract keepX Parameters for Current Combination}
\usage{
extract_keepX_for_combination(param_combinations, i, block_names)
}
\arguments{
\item{param_combinations}{Data frame of parameter combinations}

\item{i}{Row index}

\item{block_names}{Names of data blocks}
}
\value{
List of keepX values for each block
}
\description{
Extract keepX Parameters for Current Combination
}
\keyword{internal}
