% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_block_splsda.R
\name{find_best_parameters}
\alias{find_best_parameters}
\title{Find Best Parameters}
\usage{
find_best_parameters(results_df, method)
}
\arguments{
\item{results_df}{Results data frame}

\item{method}{mixOmics method used}
}
\value{
List with best parameters
}
\description{
Identifies optimal parameter combination based on performance metrics.
}
\keyword{internal}
