% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_block_splsda.R
\name{generate_param_combinations}
\alias{generate_param_combinations}
\title{Generate Parameter Combinations for Tuning}
\usage{
generate_param_combinations(ncomp, test.keepX, search_type, n_random)
}
\arguments{
\item{ncomp}{Vector of component numbers}

\item{test.keepX}{List of keepX values for each block}

\item{search_type}{Character, "grid" or "random"}

\item{n_random}{Integer, number of random combinations}
}
\value{
Data frame with parameter combinations
}
\description{
Creates all possible combinations of ncomp and keepX values for grid search,
or random sample for random search.
}
\keyword{internal}
