% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneR.r
\name{load_tunr}
\alias{load_tunr}
\title{load_tunr : selects te list of files. it is assumed that
the user has downloaded the nasa datasets and extract into your home account
however this package will use a local dataset directory with a sample set of those files}
\usage{
load_tunr(type = c("1st_test", "2nd_test"), sample_size)
}
\arguments{
\item{type}{a character string that identifies what dastaset to use , either "1st_test" or "2nd_test".}

\item{sample_size}{vibration datasets are large 984 to 2000 so this allows the user to select a sample size. if a sample_size of -1 is entered then the entire dataset is used}
}
\value{
returns a list of individual vibration files, the sample size, number of vibration channels, and datapath
}
\description{
load_tunr : selects te list of files. it is assumed that
the user has downloaded the nasa datasets and extract into your home account
however this package will use a local dataset directory with a sample set of those files
}
\examples{
Result = load_tunr("1st_test", 20 )
Result = load_tunr("2nd_test", -1 )

}
