% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneR.r
\name{machine_health_data_path}
\alias{machine_health_data_path}
\title{Title TuneR package to explore vibration datasets 
This assumes the NASA data has been extracted to the local directory}
\usage{
machine_health_data_path()
}
\value{
Sets up user specific paths
}
\description{
Title TuneR package to explore vibration datasets 
This assumes the NASA data has been extracted to the local directory
}
\examples{

}
