\name{MFCC}
\alias{MFCC}
\alias{tuneR-defunct}
\title{DEFUNCT: Mel Frequency Cepstral Coefficients}
\description{DEFUNCT: Computation of MFCCs --- this has been replaced by \code{\link{melfcc}} already and is just a wrapper! Will be removed shortly.}
\note{This function was always documented to be in development and highly EXPERIMENTAL!!!}
\seealso{\link{melfcc}}

