% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_block_splsda.R
\name{perform_block_cross_validation}
\alias{perform_block_cross_validation}
\title{Cross-Validation for Block Data}
\usage{
perform_block_cross_validation(
  X,
  Y,
  nfolds,
  fit_function,
  predict_function,
  evaluation_function,
  stratified = TRUE,
  ...
)
}
\arguments{
\item{X}{List of data matrices}

\item{Y}{Response variable}

\item{nfolds}{Number of folds}

\item{fit_function}{Function to fit model}

\item{predict_function}{Function to predict}

\item{evaluation_function}{Function to evaluate}

\item{stratified}{Whether to use stratified sampling}

\item{...}{Additional parameters}
}
\value{
CV results
}
\description{
Modified cross-validation that handles list of data blocks.
}
\keyword{internal}
