\name{plot}
\docType{methods}
\alias{plot-methods}
\alias{plot,ANY,ANY-method}
\title{S4 generic for plot}
\description{S4 generic for plot.}
\section{Methods}{
\describe{
\item{x = "ANY", y = "ANY"}{Any object for which a plot is desired.}
}}
\seealso{For the S3 generic: \code{\link[graphics]{plot.default}}}
\keyword{hplot}
\keyword{aplot}
\keyword{methods}
