% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneR.r
\name{plot_tunr}
\alias{plot_tunr}
\title{plot_tunr will produce plots for the calculated features 
so that you can explore the structure of each of the summarized features for each of the bearing measurements
looks at the response over time
These plots looks at the RMS , Kurtosis, Entropy, CrestFactor , and Skewness with parametric modelling 
the value here is to show the CI on each
 where a large CI exists there could be outliers or indication of a faulty bearing}
\usage{
plot_tunr(x_type, out_tbl)
}
\arguments{
\item{x_type}{is the freature type to be plotted, options are 
"rms", "kurtosis","entropy",",crest","skewness",mahalanobis"}

\item{out_tbl}{is the list of datasets generated by \code{\link{eda_tunr}}}
}
\description{
plot_tunr will produce plots for the calculated features 
so that you can explore the structure of each of the summarized features for each of the bearing measurements
looks at the response over time
These plots looks at the RMS , Kurtosis, Entropy, CrestFactor , and Skewness with parametric modelling 
the value here is to show the CI on each
 where a large CI exists there could be outliers or indication of a faulty bearing
}
\examples{
obj2 = eda_tunr(load_tunr("1st_test", 20 ))
plot_tunr("rms",obj2)
}
