% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune.R
\name{summary.tune_result}
\alias{summary.tune_result}
\title{Summary Method for tune_result Objects}
\usage{
\method{summary}{tune_result}(object, ...)
}
\arguments{
\item{object}{A tune_result object}

\item{...}{Additional arguments (not used)}
}
\description{
Summary Method for tune_result Objects
}
