% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{validate_cv_inputs}
\alias{validate_cv_inputs}
\title{Validate Cross-Validation Inputs}
\usage{
validate_cv_inputs(X, Y, nfolds)
}
\arguments{
\item{X}{Input data matrix}

\item{Y}{Response variable}

\item{nfolds}{Number of folds}
}
\description{
Validates inputs for cross-validation functions.
}
\keyword{internal}
