% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_block_splsda.R
\name{validate_tune_inputs}
\alias{validate_tune_inputs}
\title{Validate Tuning Inputs}
\usage{
validate_tune_inputs(X, Y, ncomp, test.keepX, search_type, method)
}
\arguments{
\item{X}{List of data matrices}

\item{Y}{Response variable}

\item{ncomp}{Vector of component numbers}

\item{test.keepX}{List of keepX values}

\item{search_type}{Search strategy}

\item{method}{mixOmics method}
}
\description{
Comprehensive input validation for tuning functions.
}
\keyword{internal}
