% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune.R
\name{validate_tune_main_inputs}
\alias{validate_tune_main_inputs}
\title{Validate Main Tune Function Inputs}
\usage{
validate_tune_main_inputs(method, data, ncomp, test.keepX, search_type)
}
\arguments{
\item{method}{Character, mixOmics method name}

\item{data}{List with X and Y components}

\item{ncomp}{Vector of component numbers}

\item{test.keepX}{List of keepX values}

\item{search_type}{Search strategy}
}
\description{
Validate Main Tune Function Inputs
}
\keyword{internal}
