\name{Arith-methods}
\docType{methods}
\alias{Arith-methods}
\alias{Arith,Wave,missing-method}
\alias{Arith,Wave,numeric-method}
\alias{Arith,Wave,Wave-method}
\alias{Arith,numeric,Wave-method}
\title{Arithmetics on Waves}
\description{Methods for arithmetics on Wave objects}
\section{Methods}{
\describe{
\item{object = "Wave"}{An object of class \code{\link{Wave}}.}
\item{object = "numeric"}{For, e.g., adding a number to the whole Wave, e.g. useful for demeaning.}
\item{object = "missing"}{For unary Wave operations.}
}}
\author{Uwe Ligges, \email{ligges@statistik.uni-dortmund.de}}
\seealso{For the S3 generic: \code{\link[base]{Arith}},  \code{\link{Wave}}}
\keyword{methods}
\keyword{arith}

