\name{Waveforms-internal}
\alias{preWaveform}
\alias{postWaveform}
\title{Internal support functions for Waveforms}
\description{Internal functions to support those for generating \link{Waveforms}.}
\usage{
preWaveform(freq, duration, from, xunit, samp.rate)
postWaveform(channel, samp.rate, bit, stereo, ...)
}
\seealso{\code{\link{Waveforms}}}
\keyword{internal}

